backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
SQLAlchemy==2.0.23
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
